extract toolbar into own file
